extract arcilator logic to tools lib
