+++ lib/CMakeLists.txt
@@ -13,4 +13,5 @@ add_subdirectory(Reduce)
 add_subdirectory(Scheduling)
 add_subdirectory(Support)
 add_subdirectory(Target)
+add_subdirectory(Tools)
 add_subdirectory(Transforms)
