+++ lib/Tools/Arcilator/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_circt_library(CIRCTArcilatorLib
+  ModelInfo.cpp
+  )
