+++ tools/arcilator/CMakeLists.txt
@@ -6,6 +6,7 @@ target_link_libraries(arcilator
   CIRCTArc
   CIRCTArcToLLVM
   CIRCTArcTransforms
+  CIRCTArcilatorLib
   CIRCTCombToArith
   CIRCTConvertToArcs
   CIRCTOM
